# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/menus": {
            "get": {
                "description": "查询菜单信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单相关"
                ],
                "summary": "查询菜单信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "current",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "新增不用传ID，修改才传ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单相关"
                ],
                "summary": "新增/修改菜单信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "传新增/修改菜单的参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateMenuReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除菜单信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单相关"
                ],
                "summary": "删除菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "删除菜单的参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.IdsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/menus/all-pages": {
            "get": {
                "description": "获取所有页面",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单相关"
                ],
                "summary": "获取所有页面",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/menus/buttons": {
            "get": {
                "description": "获取菜单的关联按钮列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单相关"
                ],
                "summary": "获取菜单的关联按钮列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 登录返回的用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/menus/constant-routes": {
            "get": {
                "description": "获取所有常量路由",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单相关"
                ],
                "summary": "获取所有常量路由",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/menus/is-route-exist": {
            "get": {
                "description": "判断路由是否存在",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单相关"
                ],
                "summary": "判断路由是否存在",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "路由名称",
                        "name": "routeName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/menus/tree": {
            "get": {
                "description": "获取菜单树",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单相关"
                ],
                "summary": "获取菜单树",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.IdsReq": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "api.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "api.UpdateMenuReq": {
            "type": "object",
            "required": [
                "menuName",
                "menuType",
                "order",
                "routeName",
                "routePath",
                "status"
            ],
            "properties": {
                "activeMenu": {
                    "description": "没有则不传",
                    "type": "string"
                },
                "component": {
                    "type": "string"
                },
                "constant": {
                    "description": "是常量(访问该路由时将不会进行登录验证和权限验证)路由则传1，否则不传",
                    "type": "boolean"
                },
                "fixedIndexInTab": {
                    "description": "固定在页签中的序号",
                    "type": "integer",
                    "maximum": 127
                },
                "hideInMenu": {
                    "description": "隐藏标签页则传1，否则不传",
                    "type": "boolean"
                },
                "href": {
                    "type": "string"
                },
                "i18nKey": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "iconType": {
                    "description": "1(iconify),2(local)",
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                },
                "id": {
                    "description": "修改才需要传，没有传算新增",
                    "type": "integer"
                },
                "keepAlive": {
                    "type": "boolean"
                },
                "menuName": {
                    "type": "string"
                },
                "menuType": {
                    "description": "1(directory),2(menu)",
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                },
                "multiTab": {
                    "description": "是多标签页则传1，否则不传",
                    "type": "boolean"
                },
                "order": {
                    "type": "integer"
                },
                "parentId": {
                    "description": "一级菜单填0",
                    "type": "integer"
                },
                "props": {
                    "description": "路由属性作跳转路由则直接传true，否则传完整json"
                },
                "query": {
                    "description": "查询条件，json传送",
                    "type": "string"
                },
                "routeName": {
                    "type": "string"
                },
                "routePath": {
                    "type": "string"
                },
                "showRole": {
                    "type": "boolean"
                },
                "status": {
                    "description": "1(enabled),2(disabled)",
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                }
            }
        }
    }
}